read can 以及项目重构
