--- can/all_in_env.txt
@@ -1,138 +1,0 @@
-absl-py @ file:///home/conda/feedstock_root/build_artifacts/absl-py_1742562417367/work
-aiofiles==24.1.0
-aiohttp==3.9.5
-aiosignal==1.3.1
-annotated-types==0.7.0
-anyio==4.4.0
-argparse-addons==0.12.0
-asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
-attrs==23.2.0
-beautifulsoup4==4.12.3
-bitstruct==8.20.0
-cantools==40.2.1
-certifi==2024.7.4
-charset-normalizer==3.3.2
-click==8.1.8
-colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
-comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1710320294760/work
-contourpy==1.2.1
-crccheck==1.3.0
-cycler==0.12.1
-dataclasses-json==0.6.7
-debugpy @ file:///D:/bld/debugpy_1719378805462/work
-decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
-diskcache==5.6.3
-distro==1.9.0
-exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1704921103267/work
-executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1698579936712/work
-fastapi==0.115.12
-ffmpy==0.5.0
-filelock==3.18.0
-fonttools==4.53.1
-frozenlist==1.4.1
-fsspec==2025.3.2
-gradio==5.29.0
-gradio_client==1.10.0
-greenlet==3.0.3
-groovy==0.1.2
-grpcio @ file:///D:/bld/grpc-split_1745189111505/work
-h11==0.14.0
-httpcore==1.0.5
-httpx==0.27.0
-huggingface-hub==0.30.2
-idna==3.7
-importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1719361860083/work
-ipykernel @ file:///D:/bld/ipykernel_1719845595208/work
-ipython @ file:///D:/bld/ipython_1719582682925/work
-jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
-Jinja2==3.1.4
-jsonpatch==1.33
-jsonpickle==3.2.2
-jsonpointer==3.0.0
-jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1716472197302/work
-jupyter_core @ file:///D:/bld/jupyter_core_1710257295326/work
-kiwisolver==1.4.5
-langchain==0.2.7
-langchain-community==0.2.7
-langchain-core==0.2.12
-langchain-experimental==0.0.62
-langchain-openai==0.1.14
-langchain-text-splitters==0.2.2
-langsmith==0.1.83
-Markdown @ file:///home/conda/feedstock_root/build_artifacts/markdown_1744984518836/work
-markdown-it-py==3.0.0
-MarkupSafe==2.1.5
-marshmallow==3.21.3
-matplotlib==3.9.1
-matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1713250518406/work
-mdurl==0.1.2
-multidict==6.0.5
-mypy-extensions==1.0.0
-narwhals==1.38.0
-neo4j==5.22.0
-nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
-networkx==3.3
-numpy==1.26.4
-openai==1.35.10
-orjson==3.10.6
-packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1718189413536/work
-pandas==2.2.3
-parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1712320355065/work
-pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
-pillow==10.4.0
-platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1715777629804/work
-plotly==6.0.1
-prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1718047967974/work
-protobuf @ file:///D:/bld/protobuf_1741124604430/work/bazel-bin/python/dist/protobuf-5.29.3-cp312-abi3-win_amd64.whl#sha256=db1130ad8784e6906680cc2b29eee1a6d9836269c954bc8c3c41186f1a93eb0a
-psutil @ file:///D:/bld/psutil_1719274712219/work
-pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
-pydantic==2.8.2
-pydantic_core==2.20.1
-pydub==0.25.1
-Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1714846767233/work
-pyparsing==3.1.2
-python-can==4.5.0
-python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1709299778482/work
-python-multipart==0.0.20
-pytz==2024.1
-pyvis==0.3.2
-pywin32==306
-PyYAML==6.0.1
-pyzmq @ file:///D:/bld/pyzmq_1715024498246/work
-regex==2024.5.15
-requests==2.32.3
-rich==14.0.0
-ruff==0.11.8
-safehttpx==0.1.6
-semantic-version==2.10.0
-setuptools==70.1.1
-shellingham==1.5.4
-six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
-sniffio==1.3.1
-soupsieve==2.5
-SQLAlchemy==2.0.31
-stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
-starlette==0.46.2
-tenacity==8.4.2
-tensorboard @ file:///home/conda/feedstock_root/build_artifacts/tensorboard_1740522445235/work/tensorboard-2.19.0-py3-none-any.whl#sha256=5e71b98663a641a7ce8a6e70b0be8e1a4c0c45d48760b076383ac4755c35b9a0
-tensorboard_data_server @ file:///D:/bld/tensorboard-data-server_1728639816333/work/tensorboard_data_server-0.7.0-py3-none-any.whl#sha256=d67bff9fae5ebc80d71e8224dbae4d34256d9c075c51d9090c5a43aa8f927346
-textparser==0.24.0
-tiktoken==0.7.0
-tomlkit==0.13.2
-tornado @ file:///D:/bld/tornado_1717722831716/work
-tqdm==4.66.4
-traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1713535121073/work
-typer==0.15.3
-typing-inspect==0.9.0
-typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1717802530399/work
-tzdata==2025.2
-urllib3==2.2.2
-uvicorn==0.34.2
-wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
-websockets==15.0.1
-Werkzeug @ file:///home/conda/feedstock_root/build_artifacts/werkzeug_1733160440960/work
-wheel==0.43.0
-wikipedia==1.4.0
-wrapt==1.17.2
-yarl==1.9.4
-zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1718013267051/work
