--- can/requirements.txt
@@ -1,11 +1,0 @@
-
-langchain>=0.1.14
-langchain-openai>=0.0.6
-langchain-community>=0.0.19
-langchain-experimental>=0.0.12
-neo4j>=5.13.0
-openai>=1.30.1
-python-dotenv>=1.0.1
-gradio>=4.26.0
-cantools>=39.3.0
-python-can>=4.3.1
