+++ can_llm/utils/project_paths.py
@@ -1,0 +1,5 @@
+import os
+
+
+def get_project_root():
+    return os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
