+++ examples/vision/val.py
@@ -1,5 +1,5 @@
 import torch
-from models.OcclusionDetectionModel import OcclusionDetectionModel
+from vision_detect.models.OcclusionDetectionModel import OcclusionDetectionModel
 from PIL import Image
 
 model_state_path = '/Volumes/My Passport/dataset/models/trained/best.pth'
