+++ setup.py
@@ -1,0 +1,12 @@
+from setuptools import setup ,find_packages
+
+setup(
+    name='FaultDetect',
+    version='',
+    packages=find_packages(),
+    url='',
+    license='',
+    author='DragonGong',
+    author_email='',
+    description=''
+)
