+++ vision_detect/utils/constant.py
@@ -1,0 +1,2 @@
+
+IMAGE_NOT_EXIST = 'The image file does not exist'
